% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{algo}
\alias{algo}
\alias{algo<-}
\alias{algo,cpss-method}
\alias{algo<-,cpss-method}
\title{Generic functions and methods: algo}
\usage{
algo(x)

algo(x) <- value

\S4method{algo}{cpss}(x)

\S4method{algo}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: algo
}
