% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{algo_param_dim}
\alias{algo_param_dim}
\alias{algo_param_dim<-}
\alias{algo_param_dim,cpss-method}
\alias{algo_param_dim<-,cpss-method}
\title{Generic functions and methods: algo_param_dim}
\usage{
algo_param_dim(x)

algo_param_dim(x) <- value

\S4method{algo_param_dim}{cpss}(x)

\S4method{algo_param_dim}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: algo_param_dim
}
