% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{cps}
\alias{cps}
\alias{cps<-}
\alias{cps,cpss-method}
\alias{cps<-,cpss-method}
\title{Generic functions and methods: cps}
\usage{
cps(x)

cps(x) <- value

\S4method{cps}{cpss}(x)

\S4method{cps}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: cps
}
