% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{cpss-class}
\alias{cpss-class}
\title{cpss: an \proglang{S4} class which collects data and information required for further change-point analyses and summaries}
\description{
cpss: an \proglang{S4} class which collects data and information required for further change-point analyses and summaries
}
\section{Slots}{

\describe{
\item{\code{dat}}{ANY.}

\item{\code{mdl}}{character.}

\item{\code{algo}}{character.}

\item{\code{algo_param_dim}}{numeric.}

\item{\code{SC}}{character.}

\item{\code{ncps}}{integer.}

\item{\code{pelt_pen}}{numeric.}

\item{\code{cps}}{numeric.}

\item{\code{params}}{list.}

\item{\code{S_vals}}{numeric.}

\item{\code{SC_vals}}{matrix.}

\item{\code{call}}{list.}

\item{\code{update_inputs}}{list.}
}}

