% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{dat}
\alias{dat}
\alias{dat<-}
\alias{dat,cpss-method}
\alias{dat<-,cpss-method}
\title{Generic functions and methods: dat}
\usage{
dat(x)

dat(x) <- value

\S4method{dat}{cpss}(x)

\S4method{dat}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}

\item{cpss}{cpss class}
}
\description{
Generic functions and methods: dat
}
