% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{mdl}
\alias{mdl}
\alias{mdl<-}
\alias{mdl,cpss-method}
\alias{mdl<-,cpss-method}
\title{Generic functions and methods: mdl}
\usage{
mdl(x)

mdl(x) <- value

\S4method{mdl}{cpss}(x)

\S4method{mdl}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: mdl
}
