% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{ncps}
\alias{ncps}
\alias{ncps<-}
\alias{ncps,cpss-method}
\alias{ncps<-,cpss-method}
\title{Generic functions and methods: ncps}
\usage{
ncps(x)

ncps(x) <- value

\S4method{ncps}{cpss}(x)

\S4method{ncps}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: ncps
}
