% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{params}
\alias{params}
\alias{params<-}
\alias{params,cpss-method}
\alias{params<-,cpss-method}
\title{Generic functions and methods: params}
\usage{
params(x)

params(x) <- value

\S4method{params}{cpss}(x)

\S4method{params}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: params
}
