% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{pelt_pen}
\alias{pelt_pen}
\alias{pelt_pen<-}
\alias{pelt_pen,cpss-method}
\alias{pelt_pen<-,cpss-method}
\title{Generic functions and methods: pelt_pen}
\usage{
pelt_pen(x)

pelt_pen(x) <- value

\S4method{pelt_pen}{cpss}(x)

\S4method{pelt_pen}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: pelt_pen
}
