% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot,cpss-method}
\alias{plot,cpss-method}
\title{plot method}
\usage{
\S4method{plot}{cpss}(obj, type, x = c(), y = c(), ...)
}
\arguments{
\item{obj}{object from cpss}

\item{type}{type of visualization}

\item{x}{x}

\item{y}{y}

\item{...}{...}

\item{cpss}{cpss class}
}
\description{
plot method
}
