% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{S_vals}
\alias{S_vals}
\alias{S_vals<-}
\alias{S_vals,cpss-method}
\alias{S_vals<-,cpss-method}
\title{Generic functions and methods: S_vals}
\usage{
S_vals(x)

S_vals(x) <- value

\S4method{S_vals}{cpss}(x)

\S4method{S_vals}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: S_vals
}
