% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SC}
\alias{SC}
\alias{SC<-}
\alias{SC,cpss-method}
\alias{SC<-,cpss-method}
\title{Generic functions and methods: SC}
\usage{
SC(x)

SC(x) <- value

\S4method{SC}{cpss}(x)

\S4method{SC}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: SC
}
