% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{SC_vals}
\alias{SC_vals}
\alias{SC_vals<-}
\alias{SC_vals,cpss-method}
\alias{SC_vals<-,cpss-method}
\title{Generic functions and methods: SC_vals}
\usage{
SC_vals(x)

SC_vals(x) <- value

\S4method{SC_vals}{cpss}(x)

\S4method{SC_vals}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: SC_vals
}
