% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary,cpss-method}
\alias{summary,cpss-method}
\title{summary method}
\usage{
\S4method{summary}{cpss}(object)
}
\arguments{
\item{object}{object from cpss}

\item{cpss}{cpss class}
}
\description{
summary method
}
