% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{tool1}
\alias{tool1}
\alias{dat}
\alias{mdl}
\alias{algo}
\alias{algo_param_dim}
\alias{SC}
\alias{ncps}
\alias{pelt_pen}
\alias{cps}
\alias{params}
\alias{S_vals}
\alias{SC_vals}
\alias{upcalle.inputs}
\title{tool1}
\usage{
dat(x)

mdl(x)

algo(x)

algo_param_dim(x)

SC(x)

ncps(x)

pelt_pen(x)

cps(x)

params(x)

S_vals(x)

SC_vals(x)

upcalle.inputs(x)
}
\arguments{
\item{x}{x}
}
\description{
tool1
}
