% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{tool2}
\alias{tool2}
\alias{dat<-}
\alias{mdl<-}
\alias{algo<-}
\alias{algo_param_dim<-}
\alias{SC<-}
\alias{ncps<-}
\alias{pelt_pen<-}
\alias{cps<-}
\alias{params<-}
\alias{S_vals<-}
\alias{SC_vals<-}
\alias{upcalle.inputs<-}
\title{tool2}
\usage{
dat(x) <- value

mdl(x) <- value

algo(x) <- value

algo_param_dim(x) <- value

SC(x) <- value

ncps(x) <- value

pelt_pen(x) <- value

cps(x) <- value

params(x) <- value

S_vals(x) <- value

SC_vals(x) <- value

upcalle.inputs(x) <- value
}
\arguments{
\item{x}{x}

\item{value}{value}
}
\description{
tool2
}
