% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{tool3}
\alias{tool3}
\alias{dat,cpss-method}
\alias{mdl,cpss-method}
\alias{algo,cpss-method}
\alias{algo_param_dim,cpss-method}
\alias{SC,cpss-method}
\alias{ncps,cpss-method}
\alias{pelt_pen,cpss-method}
\alias{cps,cpss-method}
\alias{params,cpss-method}
\alias{S_vals,cpss-method}
\alias{SC_vals,cpss-method}
\alias{upcalle.inputs,cpss-method}
\title{tool3}
\usage{
\S4method{dat}{cpss}(x)

\S4method{mdl}{cpss}(x)

\S4method{algo}{cpss}(x)

\S4method{algo_param_dim}{cpss}(x)

\S4method{SC}{cpss}(x)

\S4method{ncps}{cpss}(x)

\S4method{pelt_pen}{cpss}(x)

\S4method{cps}{cpss}(x)

\S4method{params}{cpss}(x)

\S4method{S_vals}{cpss}(x)

\S4method{SC_vals}{cpss}(x)

\S4method{upcalle.inputs}{cpss}(x)
}
\arguments{
\item{x}{x}

\item{cpss}{cpss}
}
\description{
tool3
}
