% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{tool4}
\alias{tool4}
\alias{dat<-,cpss-method}
\alias{mdl<-,cpss-method}
\alias{algo<-,cpss-method}
\alias{algo_param_dim<-,cpss-method}
\alias{SC<-,cpss-method}
\alias{ncps<-,cpss-method}
\alias{pelt_pen<-,cpss-method}
\alias{cps<-,cpss-method}
\alias{params<-,cpss-method}
\alias{S_vals<-,cpss-method}
\alias{SC_vals<-,cpss-method}
\alias{upcalle.inputs<-,cpss-method}
\title{tool4}
\usage{
\S4method{dat}{cpss}(x) <- value

\S4method{mdl}{cpss}(x) <- value

\S4method{algo}{cpss}(x) <- value

\S4method{algo_param_dim}{cpss}(x) <- value

\S4method{SC}{cpss}(x) <- value

\S4method{ncps}{cpss}(x) <- value

\S4method{pelt_pen}{cpss}(x) <- value

\S4method{cps}{cpss}(x) <- value

\S4method{params}{cpss}(x) <- value

\S4method{S_vals}{cpss}(x) <- value

\S4method{SC_vals}{cpss}(x) <- value

\S4method{upcalle.inputs}{cpss}(x) <- value
}
\arguments{
\item{x}{x}

\item{value}{value}

\item{cpss}{cpss}
}
\description{
tool4
}
