% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{update,cpss-method}
\alias{update,cpss-method}
\title{update method}
\usage{
\S4method{update}{cpss}(object, dim_update)
}
\arguments{
\item{object}{object from cpss}

\item{dim_update}{model dimension to update}

\item{cpss}{cpss class}
}
\description{
update method
}
