% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{update_inputs}
\alias{update_inputs}
\alias{update_inputs<-}
\alias{update_inputs,cpss-method}
\alias{update_inputs<-,cpss-method}
\title{Generic functions and methods: update_inputs}
\usage{
update_inputs(x)

update_inputs(x) <- value

\S4method{update_inputs}{cpss}(x)

\S4method{update_inputs}{cpss}(x) <- value
}
\arguments{
\item{x}{object from cpss}

\item{value}{value assigned to x}
}
\description{
Generic functions and methods: update_inputs
}
