% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{well}
\alias{well}
\title{Well-log data}
\format{
A vector of 4,050 measurements:
\describe{
\item{well}{Measurements.}
}
}
\source{
\doi{10.1111/1467-9868.00421}
}
\usage{
well
}
\description{
Measurements of the nuclear magnetic response of underground rocks.
}
\keyword{datasets}
